Commit of a working HTU21D library
Same base library as SparkFun's library (w attribution in source), but extended to allow for in-flight readings and better respect the spec sheet.

The "in-flight readings" are effectively just async reads; the original library was capable of this (since it used *_MEASURE_NOHOLD) but then would block on the repsonse/conversion result. There's no reason to do that; if you'd like you can just spin on isConversionComplete()